% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_table.R
\name{axis_table}
\alias{axis_table}
\title{Adds a table to a ggplot axis.}
\usage{
axis_table(.data, x = NULL, y = NULL,
  selected_cols = colnames(.data), col_widths = NULL)
}
\arguments{
\item{.data}{Data containing both the y aesthetic of the original data, as well as the columns which should be included in the table.}

\item{x}{A character vector containing the same column as is used for the x aesthetic in the plot (use only one of \code{x} and \code{y}).}

\item{y}{A character vector containing the same column as is used for the y aesthetic in the plot (use only one of \code{x} and \code{y})}

\item{selected_cols}{A character vector containing the columns in \code{.data} to be used in the axis table.}

\item{col_widths}{A numeric vector containing the desired widths (in characters) of each column.}
}
\value{
A specially-formed scale_discrete function containing the table to be added to the plot.
}
\description{
\code{axis_table}
}
\details{
Make sure you add
\code{+ theme(axis.text.x = element_text(family = "mono", color = "black", hjust=0))}

to the plot if you're adding a table to y axis or
\code{+ theme(axis.text.y = element_text(family = "mono", color = "black", angle=-90, vjust=0))}

for the x axis.
This will ensure your table is oriented correctly and that a fixed-width font is used.
}
\examples{
example to go here

}
