% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_table.R
\name{table_labeller}
\alias{table_labeller}
\title{Creates a table formatted as a ggplot labeller object.}
\usage{
table_labeller(
  .data,
  axis_col,
  selected_cols = colnames(.data),
  col_widths = NULL
)
}
\arguments{
\item{.data}{Data containing both the x/y aesthetic of the original data, as well as the columns which should be included in the table.}

\item{axis_col}{Character. The name or index of the column used for the x or y aesthetic.}

\item{selected_cols}{Character. The columns which should be used in the table.}

\item{col_widths}{Numeric vector. The desired width (in characters) of each column.}
}
\description{
\code{table_labeller} returns a function which can be used as the \code{label} parameter in \code{scale_x_discrete} or \code{scale_y_discrete}.
}
\details{
This function is not usually used directly, but is called from \code{axis_table}.
}
